deployment/requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
aiogram>=3.0.0
APScheduler>=3.10.0
lxml>=4.9.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
Pillow>=10.0.0